move model to export app
