+++ corehq/apps/accounting/models.py
@@ -331,17 +331,6 @@ class CommunicationType(object):
     )
 
 
-class ExportFileType(object):
-    CSV = "CSV"
-    EXCEL_2007_PLUS = "EXCEL_2007_PLUS"
-    EXCEL_PRE_2007 = "EXCEL_PRE_2007"
-    CHOICES = (
-        (CSV, "CSV (zip file)"),
-        (EXCEL_2007_PLUS, "Excel 2007+"),
-        (EXCEL_PRE_2007, "Excel (older versions)"),
-    )
-
-
 class Currency(models.Model):
     """
     Keeps track of the current conversion rates so that we don't have to poll the free, but rate limited API
@@ -363,29 +352,6 @@ class Currency(models.Model):
     def get_default(cls):
         default, _ = cls.objects.get_or_create(code=settings.DEFAULT_CURRENCY)
         return default
-
-
-class DefaultExportSettings(models.Model):
-    """
-    Represents the default settings for data exports
-    Currently configured via the Enterprise Settings UI
-    """
-    # Forms Exports
-    forms_filetype = models.CharField(max_length=25, default=ExportFileType.EXCEL_2007_PLUS,
-                                      choices=ExportFileType.CHOICES)
-    forms_auto_convert = models.BooleanField(default=True)
-    forms_auto_format_cells = models.BooleanField(default=False)
-    forms_include_duplicates = models.BooleanField(default=False)
-    forms_expand_checkbox = models.BooleanField(default=False)
-
-    # Cases Exports
-    cases_filetype = models.CharField(max_length=25, default=ExportFileType.EXCEL_2007_PLUS,
-                                      choices=ExportFileType.CHOICES)
-    cases_auto_convert = models.BooleanField(default=True)
-
-    # OData Forms Exports
-    odata_forms_include_duplicates = models.BooleanField(default=False)
-    odata_forms_expand_checkbox = models.BooleanField(default=False)
 
 
 DEFAULT_ACCOUNT_FORMAT = 'Account for Project %s'
@@ -457,7 +423,8 @@ class BillingAccount(ValidateModelMixin, models.Model):
     )
 
     # Default settings for data exports
-    default_export_settings = models.ForeignKey(DefaultExportSettings, null=True, on_delete=models.SET_NULL)
+    default_export_settings = models.ForeignKey('export.DefaultExportSettings', null=True,
+                                                on_delete=models.SET_NULL)
 
     class Meta(object):
         app_label = 'accounting'
